+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## [Unreleased]
 ### Added
+- Access to "edit injected content" action within language-injected snapshots
+  + For example, it allows to preview markdown/mermaid/html ...
+
+## [0.0.1]
+### Added
 - Selfie snapshot file detection and parsing
 - Language injection within snapshots
 - Automatic language injection from snapshot headers